UI prototype
Commit:
- Added components back into `~/pages/components`
    - Named with underscore prefix `_component.jsx` so Gatsby doesn't build as a page
- Updated styles on `App`
    - Added `flexbox` mixin to `index.scss`

Todo:
- People/officers section
- Client set data from json/md file or something
    - About description
    - ActionCard color, label, etc
- Routing
    - `Navbar` with `Link`
    - Learn more button with `HashRouter`
    - `ActionCard` cards with `Link`
- MOBILE???
