+++ aid/styles/ActionCard.scss
@@ -13,9 +13,6 @@
 }
 .ActionCard
 {
-    width: 520px;
-    height: 200px;
-
     background-size: cover;
     background-repeat: no-repeat;
