+++ aid/styles/index.scss
@@ -4,6 +4,14 @@
     background-image: linear-gradient($tint, $tint), url($url);
 }
 
+@mixin flexbox($dir, $wrap, $justify, $align)
+{
+    display: flex;
+    flex-flow: $dir $wrap;
+    justify-content: $justify;
+    align-items: $align;
+}
+
 
 // Global
 * {
